refactor(Main): modify Main component to accept dynamic data
The Main component was refactored to accept a `dataHorta` prop, which is an array of objects containing section data. This change allows for more flexibility and reusability by dynamically rendering sections based on the provided data. The hardcoded sections were replaced with a map function to iterate over the `dataHorta` array and render each section accordingly.